Refactor toward Reservation Card
