+++ frontend/src/app/coworking/coworking.service.ts
@@ -45,23 +45,19 @@ export class CoworkingService {
             end
         };
 
-        return this.http.post<ReservationJSON>("/api/coworking/reservation", reservation)
-            .pipe(
-                map(parseReservationJSON),
-                tap(reservation => this.status.pushReservation(reservation))
-            );
+        return this.http.post<ReservationJSON>("/api/coworking/reservation", reservation).pipe(map(parseReservationJSON));
     }
 
-    confirmReservation(reservation: Reservation) {
-        let reservationPartial = { id: reservation.id, state: "CONFIRMED" };
-        this.http.put<ReservationJSON>(`/api/coworking/reservation/${reservation.id}`, reservationPartial)
-            .subscribe((updatedReservation) => this.status.updateReservation(parseReservationJSON(updatedReservation)));
-    }
+    // confirmReservation(reservation: Reservation) {
+    //     let reservationPartial = { id: reservation.id, state: "CONFIRMED" };
+    //     this.http.put<ReservationJSON>(`/api/coworking/reservation/${reservation.id}`, reservationPartial)
+    //         .subscribe((updatedReservation) => this.status.updateReservation(parseReservationJSON(updatedReservation)));
+    // }
 
-    cancelReservation(reservation: Reservation) {
-        this.http.delete<ReservationJSON>(`/api/coworking/reservation/${reservation.id}`).subscribe((cancelledReservation => {
-            this.status.removeReservation(parseReservationJSON(cancelledReservation))
-        }));
-    }
+    // cancelReservation(reservation: Reservation) {
+    //     this.http.delete<ReservationJSON>(`/api/coworking/reservation/${reservation.id}`).subscribe((cancelledReservation => {
+    //         this.status.removeReservation(parseReservationJSON(cancelledReservation))
+    //     }));
+    // }
 
 }
