+++ frontend/src/app/coworking/reservation/reservation.component.css
@@ -1,5 +1,10 @@
+.card-container {
+    margin: 16px;
+}
+
 .mat-mdc-card {
     max-width: 640px;
+    margin: 0px;
 }
 
 .mat-mdc-card-header {
