+++ frontend/src/app/coworking/reservation/reservation.component.html
@@ -1,86 +1,5 @@
-<ng-container *ngIf="reservation$ | async as reservation">
-
-    <mat-card>
-        <!-- Card Header -->
-        <ng-container [ngSwitch]="reservation.state">
-            <mat-card-header *ngSwitchCase="'DRAFT'">
-                <mat-card-title>Confirm Reservation Details</mat-card-title>
-            </mat-card-header>
-
-            <mat-card-header *ngSwitchCase="'CONFIRMED'">
-                <mat-card-title>Reservation Confirmed</mat-card-title>
-                <mat-card-subtitle>See the XL Ambassador staffing the front desk to check-in.</mat-card-subtitle>
-            </mat-card-header>
-
-            <mat-card-header *ngSwitchCase="'CHECKED_IN'">
-                <mat-card-title>You're Checked In!</mat-card-title>
-                <mat-card-subtitle>See the XL Ambassador if you need anything. When you're done, just press the Check
-                    Out button or check out with the XL Ambassador. <br><br> Changing seats? Please check out and check
-                    back in.</mat-card-subtitle>
-            </mat-card-header>
-
-            <mat-card-header *ngSwitchCase="'CHECKED_OUT'">
-                <mat-card-title>Thanks for visiting!</mat-card-title>
-                <mat-card-subtitle>You're all checked out.</mat-card-subtitle>
-            </mat-card-header>
-
-            <mat-card-header *ngSwitchCase="'CANCELLED'">
-                <mat-card-title>Reservation Cancelled</mat-card-title>
-                <mat-card-subtitle></mat-card-subtitle>
-            </mat-card-header>
-        </ng-container>
-
-        <!-- Card Content -->
-        <mat-card-content>
-            <h3><label>Who</label> <strong>{{ reservation.users[0].first_name }} {{ reservation.users[0].last_name
-                    }}</strong></h3>
-            <mat-divider />
-
-            <h3><label>When</label> <strong>{{ reservation.start|date:'shortTime' }}</strong> until <strong>{{
-                    reservation.end|date:'shortTime' }}</strong></h3>
-
-            <p *ngIf="reservation.state == 'CONFIRMED'">Check-in between {{ reservation.start|date:'shortTime' }} and {{
-                checkinDeadline(reservation.start)|date:'shortTime' }} to avoid cancellation.
-            </p>
-            <mat-divider />
-
-            <h3><label>What</label> <strong>{{ reservation.seats[0].title }}</strong></h3>
-            <mat-divider />
-
-            <h3><label>Where</label> <strong>The XL Colab in Sitterson 156</strong></h3>
-
-            <mat-card-actions align="end" [ngSwitch]="reservation.state">
-
-                <ng-container *ngSwitchCase="'DRAFT'">
-                    <button mat-raised-button type="submit" (click)="cancel(reservation)">Cancel</button>
-                    &nbsp;
-                    <button mat-raised-button type="submit" color="primary" (click)="confirm(reservation)">{{
-                        draftConfirmationDeadline$ | async }}</button>
-                </ng-container>
-
-                <ng-container *ngSwitchCase="'CONFIRMED'">
-                    <button mat-raised-button type="submit" (click)="cancel(reservation)">Cancel</button>
-                </ng-container>
-
-                <ng-container *ngSwitchCase="'CHECKED_IN'">
-                    <button mat-raised-button type="submit" (click)="checkout(reservation)" color="primary">Check
-                        Out</button>
-                </ng-container>
-
-                <ng-container *ngSwitchCase="'CHECKED_OUT'">
-                    <button mat-raised-button type="submit" (click)="router.navigateByUrl('/coworking')"
-                        color="primary">Return to Coworking Home</button>
-                </ng-container>
-
-                <ng-container *ngSwitchCase="'CANCELLED'">
-                    <button mat-raised-button type="submit" (click)="router.navigateByUrl('/coworking')"
-                        color="primary">Return to Coworking Home</button>
-                </ng-container>
-
-            </mat-card-actions>
-
-        </mat-card-content>
-    </mat-card>
-
-
-</ng-container>
+<div class="card-container">
+    <ng-container *ngIf="reservation$ | async as reservation">
+        <coworking-reservation-card [reservation]="reservation"></coworking-reservation-card>
+    </ng-container>
+</div>
