+++ frontend/src/app/coworking/widgets/coworking-reservation-card/coworking-reservation-card.css
@@ -1,6 +1,6 @@
 .mat-mdc-card {
-    margin: 0;
     max-width: 640px;
+    margin: 0;
 }
 
 .mat-mdc-card-header {
